+++ api/app/Http/Controllers/Year2019/UnconfController.php
@@ -2,144 +2,9 @@
 
 namespace App\Http\Controllers\Year2019;
 
-use Illuminate\Http\Request;
-use App\Service\SessionService;
-use App\Http\Controllers\Controller;
-use App\Http\Controllers\ApiTrait;
+use App\Http\Controllers\BaseUnconfController;
 
-class UnconfController extends Controller
+class UnconfController extends BaseUnconfController
 {
-    use ApiTrait;
-
     protected $year = 2019;
-    protected $function = 'unconf';
-    private $missingStructure = [
-        'isBroadCast' => false,
-        'room' => [
-            'speaker_id' => 0,
-            'company' => '',
-            'company_e' => '',
-            'job_title' => '',
-            'job_title_e' => '',
-            'summary' => '',
-            'summary_e' => '',
-            'is_keynote' => false,
-            'room' => 'UnConf',
-            'floor' => '3F',
-            'img' => [
-                'web' => '',
-                'mobile' => ''
-            ],
-            'tags' => [],
-            'recordable' => true,
-            'level' => 'Basic',
-            'sponsor_id' => 0
-        ]
-    ];
-
-    /**
-     * fill correct format into json
-     */
-    public function __construct()
-    {
-        parent::__construct();
-        $this->mapMissingStructure();
-    }
-
-    /**
-     * return all unconf data.
-     *
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function index()
-    {
-        return $this->returnSuccess('success', $this->jsonAry);
-    }
-
-    /**
-     * return specific tags of unconf data.
-     *
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function getUnconfList(Request $request)
-    {
-        $result = [];
-        // loop over everyday
-        array_walk($this->jsonAry, function ($subset) use (&$result) {
-            // loop over every period
-            array_walk($subset['period'], function ($unconf) use (&$result){
-                // merge specific data in result
-                if (!empty($unconf['room'])) {
-                    $result[] = $unconf['room'][0];
-                }
-            });
-        });
-        if (empty($result)) {
-            return $this->returnNotFoundError();
-        }
-        return $this->returnSuccess('success', $result);
-    }
-
-    /**
-     * return specific id of unconf data.
-     *
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function show($id)
-    {
-        if (!is_numeric($id)) {
-            return $this->returnError('Bad request');
-        }
-        $unconfs = $this->unconfIdMap($this->jsonAry);
-        if (isset($unconfs[$id])) {
-            return $this->returnSuccess('success', $unconfs[$id]);
-        }
-        return $this->returnNotFoundError();
-    }
-
-    /**
-     * map all id and unconf conference.
-     *
-     * we can store result into redis so we don't have to do that over and over again.
-     *
-     * @return array
-     */
-    private function unconfIdMap(array $rawArr)
-    {
-        $result = [];
-        // loop over everyday
-        array_walk($rawArr, function ($subset) use (&$result) {
-            // loop over every period
-            array_walk($subset['period'], function ($unconf) use (&$result){
-                if (!empty($unconf['room'])) {
-                    $result[$unconf['room'][0]['session_id']] = $unconf['room'];
-                }
-            });
-        });
-
-        return $result;
-    }
-
-    private function mapMissingStructure()
-    {
-        // loop over everyday
-        foreach ($this->jsonAry as $key => &$subset){
-            // loop over every period
-            array_walk($subset['period'], function (&$periodSubset) {
-                $result = array_merge_recursive($this->missingStructure, $periodSubset);
-                $result['room']['started_at'] = $result['started_at'];
-                $result['room']['ended_at'] = $result['ended_at'];
-                $room = $result['room'];
-                unset($result['room']);
-                $result['room'] = [];
-                if ($result['event'] === '') {
-                    $speaker = SessionService::transportMultipleSpeaker($room);
-                    $room['speakers'] = [];
-                    $room['speakers'][] = $speaker;
-                    $result['room'][] = $room;
-                }
-                $periodSubset = $result;
-            });
-        }
-    }
 }
